+++ MovieAppApi/Controllers/WatchListItemController.cs
@@ -1,59 +1,81 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using MovieAppApi.Data;
 using MovieAppApi.DTO;
 using MovieAppApi.Models;
+using MovieAppApi.Service;
 
 namespace MovieAppApi.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class WatchListItemController : ControllerBase
     {
         private MovieContext _movieContext;
         private BuildJSON buildJSON;
+        private readonly TokenJwtService tokenJwtServ;
 
         public WatchListItemController(MovieContext movieContext)
         {
             _movieContext = movieContext;
-            this.buildJSON = new BuildJSON();
+            buildJSON = new BuildJSON();
+            tokenJwtServ = new TokenJwtService();
         }
 
-        [HttpGet("all/{watchListId}")]
-        public async Task<IActionResult> GetAll(int watchListId)
+        [HttpGet("all/{watchListId}/{userId}")]
+        public async Task<IActionResult> GetAll(int watchListId, int userId)
         {
             try
             {
-                var allWatchListItems = await _movieContext.WatchListItems
-                    .Where(w => w.WatchListId == watchListId)
-                    .ToListAsync();
-                return Ok(buildJSON.WatchListItemAll(allWatchListItems));
-            }
-            catch (Exception)
-            {
-                return BadRequest();
-            } 
-        }
-
-        [HttpPost("add")]
-        public async Task<IActionResult> AddNew([FromBody] WatchListItemDTO watchListItemDTO)
-        {
-            try
-            {
-                var oldWatchListItem = await _movieContext.WatchListItems
-                    .Where(w => w.WatchListId == watchListItemDTO.WatchListId && w.InformationMovie == watchListItemDTO.InformationMovie)
-                    .FirstOrDefaultAsync();
-                if (oldWatchListItem != null)
+                var userIdInToken = tokenJwtServ.GetUserIdFromToken(HttpContext);
+                if (int.Parse(userIdInToken) == userId)
                 {
-                    return BadRequest("Item is exist in playlist");
+                    var allWatchListItems = await _movieContext.WatchListItems
+                                        .Where(w => w.WatchListId == watchListId)
+                                        .ToListAsync();
+                    return Ok(buildJSON.WatchListItemAll(allWatchListItems));
                 }
-                var newWatchListItem = new WatchListItem();
-                newWatchListItem.WatchListId = watchListItemDTO.WatchListId;
-                newWatchListItem.InformationMovie = watchListItemDTO.InformationMovie;
-                await _movieContext.WatchListItems.AddAsync(newWatchListItem);
-                await _movieContext.SaveChangesAsync();
-                return Ok(buildJSON.WatchListItemGet(newWatchListItem));
+                else
+                {
+                    return Unauthorized();
+                }
+
+            }
+            catch (Exception)
+            {
+                return BadRequest();
+            }
+        }
+
+        [HttpPost("add/{userId}")]
+        public async Task<IActionResult> AddNew([FromBody] WatchListItemDTO watchListItemDTO, int userId)
+        {
+            try
+            {
+                var userIdInToken = tokenJwtServ.GetUserIdFromToken(HttpContext);
+                if (int.Parse(userIdInToken) == userId)
+                {
+                    var oldWatchListItem = await _movieContext.WatchListItems
+                                        .Where(w => w.WatchListId == watchListItemDTO.WatchListId && w.InformationMovie == watchListItemDTO.InformationMovie)
+                                        .FirstOrDefaultAsync();
+                    if (oldWatchListItem != null)
+                    {
+                        return BadRequest("Item is exist in playlist");
+                    }
+                    var newWatchListItem = new WatchListItem();
+                    newWatchListItem.WatchListId = watchListItemDTO.WatchListId;
+                    newWatchListItem.InformationMovie = watchListItemDTO.InformationMovie;
+                    await _movieContext.WatchListItems.AddAsync(newWatchListItem);
+                    await _movieContext.SaveChangesAsync();
+                    return Ok(buildJSON.WatchListItemGet(newWatchListItem));
+                }
+                else
+                {
+                    return Unauthorized();
+                }
+
             }
             catch (Exception)
             {
@@ -62,18 +84,27 @@ namespace MovieAppApi.Controllers
         }
 
 
-        [HttpGet("check")]
-        public async Task<IActionResult> CheckExistInWatchList([FromBody] WatchListItemDTO watchListItemDTO)
+        [HttpGet("check/{userId}")]
+        public async Task<IActionResult> CheckExistInWatchList([FromBody] WatchListItemDTO watchListItemDTO, int userId)
         {
             try
             {
-                var allWatchLists = await _movieContext.WatchListItems
-                    .Where(w => w.WatchListId == watchListItemDTO.WatchListId && w.InformationMovie == watchListItemDTO.InformationMovie)
-                    .FirstOrDefaultAsync();
-                return Ok(new
+                var userIdInToken = tokenJwtServ.GetUserIdFromToken(HttpContext);
+                if (int.Parse(userIdInToken) == userId)
                 {
-                    isExist = allWatchLists != null
-                });
+                    var allWatchLists = await _movieContext.WatchListItems
+                                        .Where(w => w.WatchListId == watchListItemDTO.WatchListId && w.InformationMovie == watchListItemDTO.InformationMovie)
+                                        .FirstOrDefaultAsync();
+                    return Ok(new
+                    {
+                        isExist = allWatchLists != null
+                    });
+                }
+                else
+                {
+                    return Unauthorized();
+                }
+
             }
             catch (Exception)
             {
@@ -106,19 +137,28 @@ namespace MovieAppApi.Controllers
         //    }
         //}
 
-        [HttpDelete("delete-one/{watchListItemId}")]
-        public async Task<IActionResult> DeleteAWatchListItem(int watchListItemId)
+        [HttpDelete("delete-one/{watchListItemId}/{userId}")]
+        public async Task<IActionResult> DeleteAWatchListItem(int watchListItemId, int userId)
         {
             try
             {
-                var watchListItemToDelete = await _movieContext.WatchListItems.Where(w => w.Id == watchListItemId).FirstOrDefaultAsync();
-                if (watchListItemToDelete == null)
+                var userIdInToken = tokenJwtServ.GetUserIdFromToken(HttpContext);
+                if (int.Parse(userIdInToken) == userId)
                 {
-                    return NotFound("Không tìm thấy mục để xóa.");
+                    var watchListItemToDelete = await _movieContext.WatchListItems.Where(w => w.Id == watchListItemId).FirstOrDefaultAsync();
+                    if (watchListItemToDelete == null)
+                    {
+                        return NotFound("Không tìm thấy mục để xóa.");
+                    }
+                    _movieContext.WatchListItems.Remove(watchListItemToDelete);
+                    await _movieContext.SaveChangesAsync();
+                    return Ok();
                 }
-                _movieContext.WatchListItems.Remove(watchListItemToDelete);
-                await _movieContext.SaveChangesAsync();
-                return Ok();
+                else
+                {
+                    return Unauthorized();
+                }
+
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -126,23 +166,32 @@ namespace MovieAppApi.Controllers
             }
         }
 
-        [HttpDelete("delete-many")]
-        public async Task<IActionResult> DeleteManyWatchListItems([FromBody] int[] watchListItemIds)
+        [HttpDelete("delete-many/{userId}")]
+        public async Task<IActionResult> DeleteManyWatchListItems([FromBody] int[] watchListItemIds, int userId)
         {
             try
             {
-                if (watchListItemIds == null || watchListItemIds.Length == 0)
+                var userIdInToken = tokenJwtServ.GetUserIdFromToken(HttpContext);
+                if (int.Parse(userIdInToken) == userId)
                 {
-                    return BadRequest("Danh sách id trống.");
+                    if (watchListItemIds == null || watchListItemIds.Length == 0)
+                    {
+                        return BadRequest("Danh sách id trống.");
+                    }
+                    var userWatchLists = await _movieContext.WatchListItems.Where(w => watchListItemIds.Contains(w.Id)).ToListAsync();
+                    if (userWatchLists.Count == 0)
+                    {
+                        return NotFound("Không tìm thấy mục để xóa.");
+                    }
+                    _movieContext.WatchListItems.RemoveRange(userWatchLists);
+                    await _movieContext.SaveChangesAsync();
+                    return Ok();
                 }
-                var userWatchLists = await _movieContext.WatchListItems.Where(w => watchListItemIds.Contains(w.Id)).ToListAsync();
-                if (userWatchLists.Count == 0)
+                else
                 {
-                    return NotFound("Không tìm thấy mục để xóa.");
+                    return Unauthorized();
                 }
-                _movieContext.WatchListItems.RemoveRange(userWatchLists);
-                await _movieContext.SaveChangesAsync();
-                return Ok();
+
             }
             catch (DbUpdateConcurrencyException)
             {
